Remove unit tests for AlternateIdentifier and Attribute components
This commit deletes all unit tests related to the AlternateIdentifier and Attribute components, including their DataMapper and Resource tests. These tests are no longer necessary, potentially due to a refactor or removal of these components. The reason behind this is that we will relay on the integration resource test for mapping and operations.